+++ EssentialFeed/EssentialFeedTests/Feed Cache/CodableFeedStoreTests.swift
@@ -210,61 +210,6 @@ class CodableFeedStoreTests: XCTestCase, FailableFeedStoreSpecs {
 		return sut
 	}
 	
-	@discardableResult
-	private func insert(_ cache: (feed: [LocalFeedImage], timestamp: Date), to sut: FeedStore) -> Error? {
-		let exp = expectation(description: "Wait for cache insertion")
-		
-		var insertionError: Error?
-		sut.insert(cache.feed, timestamp: cache.timestamp) { receivedInsertionError in
-			insertionError = receivedInsertionError
-			exp.fulfill()
-		}
-		
-		wait(for: [exp], timeout: 3.0)
-		
-		return insertionError
-	}
-	
-	@discardableResult
-	private func deleteCache(from sut: FeedStore) -> Error? {
-		let exp = expectation(description: "Wait for cache deletion")
-		var deletionError: Error?
-		sut.deleteCachedFeed { receivedDeletionError in
-			deletionError = receivedDeletionError
-			exp.fulfill()
-		}
-		wait(for: [exp], timeout: 3.0)
-		return deletionError
-	}
-	
-	private func expect(_ sut: FeedStore, toRetrieveTwice expectedResult: RetrieveCachedFeedResult, file: StaticString = #filePath, line: UInt = #line) {
-		expect(sut, toRetrieve: expectedResult, file: file, line: line)
-		expect(sut, toRetrieve: expectedResult, file: file, line: line)
-	}
-	
-	private func expect(_ sut: FeedStore, toRetrieve expectedResult: RetrieveCachedFeedResult, file: StaticString = #filePath, line: UInt = #line) {
-		let exp = expectation(description: "Wait for cache retrieval")
-
-		sut.retrieve { retrievedResult in
-			switch (expectedResult, retrievedResult) {
-			case (.empty, .empty),
-				 (.failure, .failure):
-				break
-
-			case let (.found(expectedFeed, expectedTimestamp), .found(retrievedFeed, retrievedTimestamp)):
-				XCTAssertEqual(retrievedFeed, expectedFeed, file: file, line: line)
-				XCTAssertEqual(retrievedTimestamp, expectedTimestamp, file: file, line: line)
-
-			default:
-				XCTFail("Expected to retrieve \(expectedResult), got \(retrievedResult) instead", file: file, line: line)
-			}
-
-			exp.fulfill()
-		}
-
-		wait(for: [exp], timeout: 3.0)
-	}
-	
 	private func setupEmptyStoreState() {
 		deleteStoreArtifacts()
 	}
